Extract container and update store
