+++ frontend/src/metabase/entities/groups.js
@@ -4,6 +4,7 @@ import { createEntity } from "metabase/lib/entities";
 import {
   CREATE_MEMBERSHIP,
   DELETE_MEMBERSHIP,
+  CLEAR_MEMBERSHIPS,
 } from "metabase/admin/people/events";
 import { PermissionsApi } from "metabase/services";
 
@@ -18,6 +19,8 @@ const Groups = createEntity({
   actions: {
     clearMember: async ({ id }) => {
       await PermissionsApi.clearGroupMembership({ id });
+
+      return { type: CLEAR_MEMBERSHIPS, payload: { groupId: id } };
     },
   },
 
@@ -45,6 +48,11 @@ const Groups = createEntity({
       } else {
         return state;
       }
+    }
+
+    if (type === CLEAR_MEMBERSHIPS && !error) {
+      const { groupId } = payload;
+      return assocIn(state, [groupId, "members"], []);
     }
 
     return state;
