+++ frontend/src/metabase/plugins/builtin/auth/ldap.js
@@ -8,7 +8,7 @@ import {
 
 import SettingsLdapForm from "metabase/admin/settings/components/SettingsLdapForm";
 import LdapAuthCard from "metabase/admin/settings/auth/containers/LdapAuthCard";
-import GroupMappingsWidget from "metabase/admin/settings/components/widgets/GroupMappingsWidget";
+import GroupMappingsWidget from "metabase/admin/settings/containers/GroupMappingsWidget";
 
 PLUGIN_ADMIN_SETTINGS_UPDATES.push(
   sections =>
